+++ backend/src/test/java/com/ice/studyroom/domain/membership/infrastructure/email/EmailVerificationServiceTest.java
@@ -12,7 +12,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.ice.studyroom.domain.membership.application.EmailVerificationService;
+import com.ice.studyroom.domain.identity.infrastructure.email.EmailVerificationService;
 import com.ice.studyroom.global.exception.BusinessException;
 import com.ice.studyroom.global.service.CacheService;
 import com.ice.studyroom.global.service.EmailService;
